Huge Updates
Used a function to be able to run the code over different date ranges
Organised the outputs so that they are arranged as requested - four CSV's with the differeent ranges on each.
Arranged code so that the functions are all at the top and then the code runs through them at the bottom.
Changes to config file to suit.